requirements.txt
Django==4.0.2
django-environ
factory-boy==3.2.1
